refactor: Minor fixes and renaming namespace
