+++ penna-yaml-config/src/test/java/penna/config/yaml/parser/ParserTests.java
@@ -4,6 +4,10 @@ import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junitpioneer.jupiter.cartesian.CartesianTest;
+import penna.config.yaml.parsers.JacksonParser;
+import penna.config.yaml.parsers.Parser;
+import penna.config.yaml.parsers.SnakeyamlEngineParser;
+import penna.config.yaml.parsers.SnakeyamlParser;
 
 import java.io.IOException;
 import java.nio.file.Files;
